rename new connection button, revert agentbase file location
